Extract binary specific model server runner implementation to binary_kinds.
BinaryKind is an abstraction about what kind of serving binary we're using during the infra validation. Many setup configuration depends on the kind of binary:

- In model server:
  - What docker image should it be running
  - What environment variables should be set
  - Which port is used
  - (in docker) Volume mounts
- In model server client:
  - What kind of request should it send

and those are what BinaryKind is trying to absract.

Note: Origianl local docker runner had tensorflow serving specific implementation that needs to be decoupled away. This also makes tensorflow serving implementation reusable in kubernetes runner.

Also renamed LocalDockerModelServerRunner to LocalDockerRunner, as it is shorter but precise enough.

PiperOrigin-RevId: 294602049