Refactor model server runner and model server client interface.
Now model server runner is only dealing with the infra to launch the instance and wait until it is up and running. It does not care how the model server actually works. Previously model server runner implementation was tightly coupled with model server client, thus leading to poor abstraction.

As we have clear boundary between infrastructure (i.e. dealing with model server instance) and the service (i.e. dealing with model server service), we can classify each error cases explicitly, which are expressed as error_types.

Although we could distinguish the service error from the infrastructure error, we're regarding every exception the same, and retry the whole infra validation workflow up to given number. (As opposed to only retrying when infrastructure has failed) This is more of a philosophical decision to regard the whole workflow as a black box, assuming nothing about the internal structure, while we can clearly have benefit with various error types as user can understand the issue more clearly.

PiperOrigin-RevId: 295132712